savings count

+++ app/dto/dasboard_response_dto.py
@@ -32,7 +32,7 @@ class OverallProgressDTO(BaseModel):
 
 class SavingsDTO(BaseModel):
     savingsPercentage: float
-    overallSavingsAmount: int
+    overallSavingsAmount: float
     pdaSavings: int
     fdaSavings: int
     percentage_savings_fda: float
